chatbot/requirements.txt
streamlit
openai
llama-index
nltk
python-dotenv
